Move files to respect new folder structure
